getNextAutoIncrement long to int
This method was changed to return int. Previously was long. Reason was
that the databases were anyway treating the value as int and thus there
was no point having to additionally convert the value in long.

Also there are chances that later int will be converted to string
anyways because of issue #1