TitanEmbedded set & get user success
The titan embedded successfully implemented the set and get user methods
defined by databaseService.

The test file was updated to have the following:
1. Junit 4
2. BeforeClass and AfterClass functions for one time setup.
3. Test case updates to include all test cases in the implementation
class.
4.Impl class now has a better transaction handling.

This commit also fixes issue #3.
Index was not getting recogized because the propertyName under property
key was wrong. Corrected it to show nodeId instead of nodeIdPropertyKey
and userId instead of userIdPropertyKey.