Bug fix: Filesystem Serive restore filesystem
Scenario: Filesystem was there. Then it got deleted. Then a new
filesystem got created with the same name. Then the deleted filesystem
was restored. End result would be with two filesystems of the same name.

Fix: Before restoring a filesystem, checks if the same name filesystem
is already present. If yes, throws error.

Workaround for user: Later there will be a function to rename the
filesystem/file/directory so the new entity can be renamed and then the
deleted entity can be bought back.