Commit Operation
Previous commit to addCommit() operation support was rolled back because
the operation to commit a bunch of files together will be handled
differently than what I thought before.

How bunch commit will work now:
Every time a bunch commit is requested, a new version of filesystem will
be created. The old filesystem will be copied and the new copied
filesystem will have all the changes. This will cause massive
redundancy, but currently I dont have any other good methods to achieve
this.

Problem Statement:
To have a bunch commit system like github where you can commit a bunch
of changes at once. Then at a later time you can rollback those changes
and expect the filesystem to be in the exact same state when that change
was made. To be able to restore the exact same state, its necessary to
store the state of all the files/dirs at that moment the change was
made.

Solution 1:
A new commit node for each commit and all the files and directories
point to that commit which belongs to it. Those that didn't change must
point to a previous commit. This way each file will belong to some
commit. To revert back to the previous state, start with the most latest
commit and start rolling back each operation that happened.
Issue: Many links such as has->had and vice versa are getting changed
and keeping track of them is messy. Also, so many relations to a node
and having to maintain all these introduce complications. Also, to
revert back a hundred commits, it would take huge time as the revert
needs to happen sequentially from most recent to least recent.

Solution 2:
No concept of commit. Let the users manually make a snapshot of the
system and they can return to that snapshot anytime.
Issue: What if users forget to make a snapshot? Also, had this just been
a filesystem version system, this solution would make sense, however
this is also a code version system. In code, things are constantly
changing, and a github like bunch commit option is required.

Solution 3:
This solution is the one currently been implemented. For al bunch
commits, create a new fileysstem and copy all the old nodes and make the
new changes. So the new filesystem will be old filesystem + new changes.
To revert back to any filesystem is a unit operation time because the
graph needs not be re-calculated. The graph is already present there and
can be restored in unit time.
Issue: Data redundancy. If there are 100 nodes and only one of them
change. To reflect that one change, 99 nodes are copied (redundant), and
one node has the change.

Solution 4:
A better way for the above solution. Only keep the hierarchy of the
changed node. An absense of hierarchy indicates to get that change from
the previous filesystem. This way only those nodes gets recorded which
have changed and hence no redundancy.
Issue: We lose the unit time capability to restore. To restore, the
graph needs to be restored first. To restore graph, start from the most
recent graph and union it with the previous version. Keep doing this
until the desired graph is reached. This union operation will take time.
This can be later implemented if need be.


Other Changes:
===============

Version property moved to enum

The version property was moved to enum as opposed to having a string in
the code.