Commit Id
With this commit, each directory/file operation which changes the graph
will have a commitId on the new relationship. This commitId in a graph
will uniquely determine the new changes in the new filesystem.

Scenario:
A file gets deleted. Another file with same name is created and deleted.
Now there will be two files with the "had" relationship and when that
file is trying to get restored, the system will have no way of
distinguishing which file to restore. The problem is the primary
assumption that combination of path and name is unique. But when things
get deleted, this assumption is no longer true. A commitId solves this
issue because each relationship will now have a commitId associated with
them. A combination of all three will be unique always and hence the
user can specify which one to restore.