Package changes default packages chagned to bookeeping.backend.*
It has been realized later in the development that the rest api for
bookeeping will not be a part of bookeeping database api. Reasons are
since bookeeping database is likely to change a lot of time, but the
rest api will be pretty much straightforward and hence will not go as
much change as the former one. Having two separate projects and the
latter project using the former project as a jar, it makes clear
separatation between the team managing the rest api and the team
managing the backend part. Also the documentation for backend can be
written separately for "developers" and docs for rest can be written on
a general scale for "developers" and for "end users".