Filesystem Create Removed from Commit
The filesystem create operation is removed from the bounds of commit
operation. This should be like a standalone operation just like user
create because this is the parent container under which commits will be
present. There is no reason for this operation to be in the commit
itself.

Also its logical to state that if a commit is happenning in a
filesystem, that filesystem MUST exist. Hence filesystem must be out of
the commit bounds.